+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * The plugin provides an ESM, CSJ and UMD version
 * The "class" option accepts a single class name or an array of class names
 * Sources were converted to TypeScript
-* replace deprecated properties "page{X,Y}Offset" with scroll{X,Y}
+* Replace deprecated properties "page{X,Y}Offset" with scroll{X,Y}
+* Dropped support for "currency" and "currencyFormatCallback" options (use "tooltipFnc" instead)
 
 ## [1.0.0] - 2nd January 2023